Tyler delete

+++ PcSales/Models/MoboSpec.cs
@@ -9,7 +9,7 @@ namespace PcSales.Models
         public string FormFactor { get; set; }
         public int SataSlots { get; set; }
         public string Socket { get; set; }
-        public double MemoryMax { get; set; }
+        public int MemoryMax { get; set; }
         public string MemoryType { get; set; }
         public int MemorySlots { get; set; }
         public int M2Slots { get; set; }
